weibull start values
The uniroot function is used to calculate the method of moments start
values. In some case there is not a real valued solution, ie the
function doesn't cross the horizontal axis, in that case the minimum
value is used. The is not technically correct  but will suffice for
starting values for the maximum likelihood routine.