Organizing files
All of the R files in code/function should be functions that can be
sourced without any actual evaluation happening. Each R file name
should match exactly the name of the function contained and only one
function should be in a file. The R files in code/
are for implementing the functions on the data.